It's been a mess ...
in commit messages and they way I did the changes, so I decided to just
rebase it all forewer, hide my stupidity, and to be honest, I'm not
proud of it. I'll try my best in the future.

For now I have to admit that I did the force push after a huge rebase.